+++ tools/gcb.bazelrc
@@ -1,6 +1,7 @@
 build --strategy=KotlinCompile=worker
 build --remote_cache=https://storage.googleapis.com/arcs-github-gcb-bazel-cache --google_default_credentials
 build --verbose_failures --spawn_strategy=sandboxed --jobs=32 --ram_utilization_factor=33
-build --nostamp
+build --nostamp --nodesugar_for_android
 test --spawn_strategy=sandboxed --nocache_test_results --test_output=errors --jobs=32 --ram_utilization_factor=33
-test --nostamp
+test --nostamp --nodesugar_for_android
+
