+++ tools/local-presubmit
@@ -18,7 +18,8 @@ status "Testing all of the things."
 set -x
 
 ./tools/sigh webpack
-./tools/bazelisk ${BAZELRC_OPTS} test //java/... //javatests/... //src/... //particles/...
+./tools/bazelisk ${BAZELRC_OPTS} test --build_tests_only -- //java/... //javatests/... //src/... //particles/... -java/arcs/sdk/android/dev/... -javatests/arcs/sdk/android/dev/...
+#./tools/bazelisk ${BAZELRC_OPTS} --desugar_for_android test -- //java/arcs/sdk/android/dev/... //javatests/arcs/sdk/android/dev/...
 ./tools/sigh testShells
 ./tools/sigh lint
 ./tools/sigh test --all
